requirements.txt
# Text-to-SQL Neo4j Graph RAG Dependencies

# Core
pyyaml>=6.0
python-dotenv>=1.0.0

# Neo4j
neo4j>=5.0.0

# OpenAI
openai>=1.0.0

# LlamaIndex (optional, for advanced features)
# llama-index>=0.10.0
# llama-index-graph-stores-neo4j>=0.1.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
